create branch of leticia and frist commit

+++ automatização/src/main/java/com/rh_ponto/automatizacao/dto/ColaboradorDTO.java
@@ -1,15 +1,11 @@
 package com.rh_ponto.automatizacao.dto;
 
-
 import lombok.Data;
 
 @Data
 public class ColaboradorDTO {
-    private Long idColaborador;
+    private Integer idColaborador;
     private String nome;
     private String cpf;
-    private Long setorId;
-
-
-
+    private Integer setorId;
 }
