+++ automatização/src/main/java/com/rh_ponto/automatizacao/dto/SetorDTO.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class SetorDTO {
-    private Long idSetor;
+    private Integer idSetor;
     private String nomeSetor;
 }
